requirements.txt
=Django==3.2
django-phonenumber-field==6.0.0
phonenumbers==8.12.41

